Create README.md
«README: Intro to the project
Hi, I'm <Username> and I'm an avid learner of computer science and programming. I love exploring new and exciting technologies and applying them in different projects.
I'm proud to say that I've been working on some projects in GitHub and I'm excited to explore more ideas and technologies in the future.
As they say "the sky's the limit", and I can't wait to see what other projects I can create and explore in the future!
»